feat: 챌린지 인증 포인트 미지급 기능
