+++ src/main/java/com/example/green/domain/certification/domain/ChallengeCertification.java
@@ -80,11 +80,18 @@ public class ChallengeCertification extends TimeBaseEntity {
 		validateNullData(certifiedDate, "인증 시각은 필수값입니다.");
 	}
 
+	public boolean canApprove() {
+		return status != CertificationStatus.APPROVED;
+	}
+
 	public void approve() {
 		this.status = CertificationStatus.APPROVED;
 	}
 
 	public void reject() {
+		if (status == CertificationStatus.APPROVED) {
+			throw new CertificationException(CertificationExceptionMessage.ALREADY_APPROVED_CERT);
+		}
 		this.status = CertificationStatus.REJECTED;
 	}
 }
