+++ src/main/java/com/example/green/domain/certification/ui/dto/CertificationIdentifierDto.java
@@ -5,5 +5,5 @@ import java.util.List;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 
-public record CertificationVerifyDto(@NotNull @NotEmpty List<Long> certificationIds) {
+public record CertificationIdentifierDto(@NotNull @NotEmpty List<Long> certificationIds) {
 }
