+++ src/main/java/com/example/green/domain/certification/util/CertificationClientHelper.java
@@ -1,17 +1,22 @@
 package com.example.green.domain.certification.util;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
+import java.util.List;
 
 import org.springframework.stereotype.Component;
 
+import com.example.green.domain.certification.domain.ChallengeCertification;
 import com.example.green.domain.certification.domain.ChallengeSnapshot;
 import com.example.green.domain.certification.domain.MemberSnapshot;
 import com.example.green.domain.common.service.FileManager;
 import com.example.green.global.client.ChallengeClient;
 import com.example.green.global.client.MemberClient;
+import com.example.green.global.client.PointClient;
 import com.example.green.global.client.dto.ChallengeDto;
 import com.example.green.global.client.dto.ChallengeGroupDto;
 import com.example.green.global.client.dto.MemberDto;
+import com.example.green.global.client.request.PointEarnRequest;
 
 import lombok.RequiredArgsConstructor;
 
@@ -22,6 +27,7 @@ public class CertificationClientHelper {
 	private final ChallengeClient challengeClient;
 	private final MemberClient memberClient;
 	private final FileManager fileManager;
+	private final PointClient pointClient;
 
 	public ChallengeSnapshot getPersonalSnapshot(Long challengeId, Long memberId, LocalDate challengeDate) {
 		ChallengeDto dto = challengeClient.getPersonalChallengeByMemberAndDate(challengeId, memberId, challengeDate);
@@ -44,5 +50,17 @@ public class CertificationClientHelper {
 
 	public void processCertSideEffect(String imageUrl) {
 		fileManager.confirmUsingImage(imageUrl);
+	}
+
+	public void processApproveSideEffect(List<ChallengeCertification> certs) {
+		List<PointEarnRequest> request = certs.stream()
+			.map(cert -> new PointEarnRequest(
+				cert.getMember().getMemberId(),
+				BigDecimal.valueOf(cert.getChallenge().getChallengePoint()),
+				cert.getChallenge().getChallengeId(),
+				cert.getChallenge().getChallengeName()
+			))
+			.toList();
+		pointClient.earnPoints(request);
 	}
 }
