+++ src/main/java/com/example/green/domain/challenge/util/ClientHelper.java
@@ -1,17 +1,13 @@
 package com.example.green.domain.challenge.util;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Component;
 
-import com.example.green.domain.certification.domain.ChallengeCertification;
 import com.example.green.global.client.MemberClient;
-import com.example.green.global.client.PointClient;
 import com.example.green.global.client.dto.MemberDto;
-import com.example.green.global.client.request.PointEarnRequest;
 
 import lombok.RequiredArgsConstructor;
 
@@ -20,7 +16,6 @@ import lombok.RequiredArgsConstructor;
 public class ClientHelper {
 
 	private final MemberClient memberClient;
-	private final PointClient pointClient;
 
 	public String requestMemberKey(Long memberId) {
 		return memberClient.getMember(memberId).getMemberKey();
@@ -38,17 +33,5 @@ public class ClientHelper {
 			.entrySet()
 			.stream()
 			.collect(Collectors.toMap(Map.Entry::getKey, entry -> entry.getValue().getMemberKey()));
-	}
-
-	public void processApproveSideEffect(List<ChallengeCertification> certs) {
-		List<PointEarnRequest> request = certs.stream()
-			.map(cert -> new PointEarnRequest(
-				cert.getMember().getMemberId(),
-				BigDecimal.valueOf(cert.getChallenge().getChallengePoint()),
-				cert.getChallenge().getChallengeId(),
-				cert.getChallenge().getChallengeName()
-			))
-			.toList();
-		pointClient.earnPoints(request);
 	}
 }
